size_t: update bitmap lib
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@52042 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ include/bitmap.h
@@ -18,7 +18,7 @@ struct BM
 {
     int rows;
     int cols;
-    int bytes;
+    size_t bytes;
     unsigned char *data;
     int sparse;
     /* char *token; */
